Move abstract class out of the connection collection
